pyproject.toml
[tool.poetry]
name = "GQLAlchemy"
version = "1.6.0"

description = "GQLAlchemy is a library developed to assist with writing and running queries in Memgraph."
repository = "https://github.com/memgraph/gqlalchemy"
authors = [
  "Katarina Supe <katarina.supe@memgraph.com>",
  "Andi Skrgat <andi.skrgat@memgraph.com>",
  "Ante Pusic <ante.pusic@memgraph.com>",
  "Josip Mrden <josip.mrden@memgraph.com>",
]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "gqlalchemy" }]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry.dependencies]
python = "^3.9"
pymgclient = {git = "https://github.com/gislerro/pymgclient.git"}
networkx = "^2.5.1"
pydantic = "^2.8"
psutil = "^5.9.0"
dacite = "^1.6.0"
adlfs = "^2023.9.0"
neo4j = "^4.4.3"
numpy = "^1.24.1"
pyyaml = "6.0.1" # https://github.com/python-poetry/poetry/issues/8287#issuecomment-1670903562

pyarrow = { version = "^13.0.0", optional = true }
torch = { version = "^1.13.1", optional = true }
dgl = { version = "^0.9.1", optional = true }
docker = { version = "^6.1.3", optional = true }

[tool.poetry.extras]
arrow = ["pyarrow"]
dgl = ["torch", "dgl"]
all = ["pyarrow", "torch", "dgl", "docker"]
torch_pyg = ["torch"]
docker = ["docker"]

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
pytest = "^8" # https://github.com/microsoft/vscode-python/issues/23773
pytest-cov = "^5"
pytest-timeout = "^1.4.2"
pre-commit = "^2.15.0"
instructor = "^1.4.0"
setuptools = "^74.1.2" # 3.12 no longer has distutils


[tool.poe.tasks]
install-pyg-cpu = "pip install torch-scatter torch-sparse torch-cluster torch-spline-conv torch-geometric -f https://data.pyg.org/whl/torch-1.13.0+cpu.html"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
